pyproject.toml
[project]
name = "test-psek"
version = "0.1.0"
description = ""
authors = [
    {name = "colosss",email = "kocnevs.jk@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standart] (>=0.135.3,<0.136.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose[cryptograghy] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"
pytest-cov = "^5.0.0"